Add ToggleControl component to player
The ToggleControl allows for multiple components to be enabled/disabled at the same time.

It takes all "ToggleableBehaviour" classes and then enables/disables all of them when the ToggleControl component is enabled or disabled, which can be done from any script by calling the Toggle function. Currently, the PlayerController, PunchHandler, and Rage classes are all ToggleableBehaviour because they involve player input